requirements.txt
diffusers
torch
accelerate
Pillow
gradio